backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
pyjwt==2.10.1
python-dotenv==1.0.0
bcrypt==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0
greenlet==3.0.1
email-validator==2.1.0.post1
asyncpg==0.29.0
psycopg2-binary==2.9.9
